Add support for Siglent eCal emulation.
This patch allows a Siglent SVA1032X to calibrate using a LibreCAL.  It
adds support in the firmware to boot into a Siglent-compatible mode
(removes USB MSC, to avoid the instrument's UI trying to write to the
disk; adds a very limited USB TMC that knows how to speak the Siglent
calibration commands) when Siglent calibration coefficients are present,
and the unit has the FUNCTION key held while powering on.

We also add a `convert_siglent.py` script, which converts the
calibration data stored in the LibreCAL's native Touchstone format to
the reverse-engineered header and ZIP+CSV format that a Siglent VNA
requires.  Some fields in the Siglent header are not fully understood,
but they don't seem to make a difference at least on SVA1032X.

The calibration appears to provide good results tested against other
calibration standards I have lying around, and so the S-parameter
conversion for ports 1 and 2 at least are probably correct.
Unimplemented is an "ATT" mode, used for the VNA's Confidence Check
function; I suspect this may be an attenuator, but without a real
Siglent eCal, I can't know for sure, and more to the point, the LibreCAL
doesn't have one other than the Through port, so we do not implement it.

Signed-off-by: Joshua Wise <joshua@joshuawise.com>